+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add `blackboard categorize` command from `krerkkiat/ta-tooling`.
 - Add `github get-team-id` command from `OU-CS3560/org-bulk-invite`.
+- Fix not being able to `Ctrl-C` the `blackboard student-list` command on Windows.
 
 ## v0.2.0